武器ごとに攻撃が変わるようになった。

+++ Assets/Data/dictinary/List/WeaponDataList.cs
@@ -250,7 +250,7 @@ public struct WeaponData
     /// <summary>
     /// UŒ‚—Í
     /// </summary>
-    public float AttackPoint;
+    public int AttackPoint;
 
     /// <summary>
     /// •Šíí
